Wave 38: 3431 -> 3497 (+66), bytes 47.35% -> 48.70%
66 functions promoted, verified by exit code across 12 agents (3 concurrent,
the user's cap). Both builds reproduce SHA1
14dd0b22c894865867aff89e8116b2dffae25605: `make SPLIT=1 compare` and
`make compare`, run and read separately.

THE LOOP BAND IS OPEN. Wave 37 measured it at 39/41 and told wave 38 to batch
there; overlap_screen.py was still hard-filtering `backward_branches == 0`, so
its batching list read 51 blocks / 71 candidates / 11KB. `--allow-loops` lifts
it and the same screen returns 96 blocks / 496 candidates / 62KB. The plain run
now self-reports what the filter hides (532 functions, 66KB in band), mirroring
the --min-size and --block-max notices. 51 of this wave's 66 carry a backward
branch; the two best batches of the wave (7/7 and 6/6) were loop-heavy blocks.
Fourth screen artifact found in this one tool and the largest.

A THIRD DUPLICATE TIER: CALLEE-NORMALISED. The strict and loose tiers both key
on the literal callee name, so two functions that are ONE SHAPE calling a
different sibling never group. Found by hand while batching: sub_08028A68 and
sub_08028AEC are 132 bytes and 56 instructions each and differ in exactly four
`bl` sites naming sub_08028894 against sub_08028874. `bl` targets are now
indexed by order of first appearance -- so two functions still only group if
they call their callees in the same PATTERN -- and it reports ten pairs
ROM-wide, all invisible to both older tiers. Confirmed against a known result
(sub_08071CC4/sub_08071CDC, matched together by W38-A before the tier existed)
and operationally (both A68/AEC and the 200-byte 0x08041 pair matched, one
derivation each). It also surfaces sub_08070544/sub_08070578, two PARKED
functions that are one shape.

Checking a lead is worth as much as having it. sub_080415E4/sub_080416A4 look
like twins by size, address, block and branch count; they are 84 instructions
against 81. The agent was told so explicitly and matched both independently --
the same false-positive mode waves 21 and 24 recorded for size signatures.

M4A NEEDS old_agbcc SELECTIVELY, AND THE DISCRIMINATOR IS A CALL IN A LOOP.
The brief predicted all five MPlay*Control entry points would need the
override. W38-A refuted it: sub_08071420/88/4FC match under BOTH compilers,
sub_08071584/5F8 need old_agbcc, and those two are exactly the ones whose loop
body contains a `bl` -- where r0 is both the argument register and the only
free scratch. Both misses were 94.8% with the SIZE ALREADY EXACT. It then
tested the converse with throwaway entries (all five match under old_agbcc, so
it is a strict superset here) and REMOVED them again: an override that changes
no bytes is unfalsifiable and stops the file being evidence.

Three harness bugs, each of which produced a wrong-looking ROM rather than an
error message:

  - A target-specific `CC1 :=` does not invalidate an existing .o. New override
    entries + regenerated .mk + rebuild = a 12-byte ROM mismatch inside exactly
    the two overridden functions, which were simply never recompiled. The C
    object rule now depends on build/compiler-overrides.mk, and
    gen_overrides_mk.py writes only when the CONTENT changes so an ordinary
    wave (promoted.json moves every time) triggers no rebuild.
  - gen_overrides_mk.py refused ANY unit sharing, including the case where every
    function in it carries the SAME override -- where the hazard cannot occur.
    promote.py had merged the two old_agbcc functions into one unit, so a
    correct pair was unbuildable and the advice on offer was to fight promote's
    own layout. It now compares the FLAGS (cc1/cflags), not whole entries:
    `reason` is prose and differs per function by design, which made two
    identically-configured functions look like a conflict.
  - An agent retyped promoted sub_0804247C (int -> bool8, sound: one `bl` site
    ROM-wide, in sub_08041758, zero-extending with `lsls #0x18; lsrs #0x18`,
    body returns literals so nothing outranks the call site) and DELETED its
    draft. trymatch compiles the draft, so the oracle went blind on a file that
    still compiles into the ROM. Caught only by the final from-scratch sweep --
    it had passed the mid-wave regression sweep an hour earlier.

"Callers convert implicitly" is a claim about what callers PASS, not what they
could pass. W38-A retyped sub_08071420/sub_08071488's first parameter to
`struct MusicPlayerInfo *` noting callers passing `void *` convert implicitly.
True -- but the callers pass `extern u8 gUnknown_03005BA0[]`, and `u8 *` does
not convert. Broke the split build in three separate files, one build cycle
each. Settled by retyping all EIGHT objects of that group (the wave-24 and
wave-32 notes both said the real type was MusicPlayerInfo and declared `u8 []`
because no member was reachable from C -- W38-A's match is the condition they
were waiting on). Kept as `[]` so array decay leaves every call site untouched;
all eight re-verified byte-for-byte.

Also: c_08075E68.c defined its own `struct Unk8074834Out` for sub_08074834's
out-parameter, byte-identical in layout to the `struct Unk0202FE38` W38-B
identified. One object, two tag names; the local is gone.

The wave-37 inherited-draft rule held, and narrowed. sub_0807519C carried a
97.7% draft, was rewritten from scratch, and matched. sub_08042864 carried a
98.57% draft, was rewritten from scratch on the same instruction, and landed at
98.6% -- so a high score means the cheap paths are gone, and a rewrite recovers
nothing when the cheap paths were never the problem. Parked with that reading.

Parked 4 with exact diffs and ruled-out axes: sub_08074AD0 (-4B, the pool word
belongs mid-function), sub_0803E6C4 (-4B, fresh local vs reused parameter),
sub_0801F2AC (100.0%, allocation-only, and its work/ draft is a LATER WORSE
attempt than its best.c), sub_08042864. data/parked.json had drifted to
indent=2/LF; normalised back to indent=1/CRLF in the same commit.

+++ src/decomp/c_08039064.c
@@ -7,18 +7,7 @@
  * sub_08039064 @ 0x08039064, sub_080390CC @ 0x080390CC
  */
 
-/* A four-way comparison of two parallel s8 tables at gUnknown_0849D5F8, run on
- * the pair (n - 1, n): the unk20 table decides first and the unk2c table only
- * breaks its tie, so this reads a direction/ordering between adjacent slots.
- *
- * IT FALLS OFF THE END DELIBERATELY. When both pairs compare equal no `movs r0`
- * executes and r0 still holds the raw parameter -- the ROM has no default arm
- * and agbcc emits none. Adding `return 0;` would cost a `movs r0,#0`.
- *
- * The `ldrb` + `lsls #24; asrs #24` on an `s8` member is not a cast in the
- * source: the index is a register, so the address is materialised first and
- * the signed byte load loses its register-offset `ldrsb` form. */
-int sub_08039064(u8 n)
+u8 sub_08039064(u8 n)
 {
     struct Unk0849D5F8 *p;
 
@@ -37,15 +26,7 @@ int sub_08039064(u8 n)
         return 2;
 }
 
-/* sub_08039064's ascending twin: the same two parallel s8 tables compared on
- * the pair (n, n + 1) instead of (n - 1, n), with the four results permuted
- * (1/3/2/4 against 3/1/4/2). The guard is the table's own extent at unk45
- * rather than a zero test, which is what makes this the top-of-range case.
- *
- * The base lands in ip because the guard's compare needs the pointer live
- * across it while r0..r3 are all busy; it is register pressure, not a loop.
- * Falls off the end on the all-equal path exactly as the twin does. */
-int sub_080390CC(u8 n)
+u8 sub_080390CC(u8 n)
 {
     struct Unk0849D5F8 *p;
